시간 바 UI 적용
- 영업 시작, 종료 마커 위치는 Content/Blueprints/General/BP_NormalBakeryGameMode의 영업 시간으로 조정

+++ Source/Bakery/Private/General/BakeryGameState.cpp
@@ -62,7 +62,7 @@ void ABakeryGameState::AddDay()
 	AddDay(1);
 }
 
-void ABakeryGameState::SetTime(int Minute)
+void ABakeryGameState::SetTime(float Minute)
 {
 	ElapsedTime = Minute;
